pyproject.toml
[project]
name = "translator-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain-google-genai>=2.1.12",
    "langchain-groq>=0.3.8",
    "python-dotenv>=1.2.1",
]
